refactor: group type definitions under src/types
