+++ src/perf.ts
@@ -1,23 +1,16 @@
-export interface PerformanceEntryLike {
-  duration: number;
-}
+import type {
+  PerformanceEntryLike,
+  PerformanceLike,
+  PerformanceTimer,
+  TimedPerformance,
+} from './types/perf';
 
-export interface PerformanceLike {
-  now(): number;
-  mark(name: string): void;
-  measure(name: string, startMark?: string, endMark?: string): void;
-  clearMarks(name?: string): void;
-  clearMeasures(name?: string): void;
-  getEntriesByName(name: string, type?: string): PerformanceEntryLike[];
-}
-
-export interface PerformanceTimer {
-  stop(): number;
-}
-
-export interface TimedPerformance extends PerformanceLike {
-  start(label: string): PerformanceTimer;
-}
+export type {
+  PerformanceEntryLike,
+  PerformanceLike,
+  PerformanceTimer,
+  TimedPerformance,
+} from './types/perf';
 
 let cached: Promise<TimedPerformance> | undefined;
